fix rotation, move priority skeleton update to WebGLRendererSystem
