+++ packages/engine/src/renderer/WebGLRendererSystem.ts
@@ -19,6 +19,8 @@ import {
   PCFSoftShadowMap,
   PerspectiveCamera,
   ShadowMapType,
+  Skeleton,
+  SkinnedMesh,
   sRGBEncoding,
   ToneMapping,
   WebGL1Renderer,
@@ -45,6 +47,8 @@ import { overrideOnBeforeCompile } from '../common/functions/OnBeforeCompilePlug
 import { Engine } from '../ecs/classes/Engine'
 import { EngineActions, getEngineState } from '../ecs/classes/EngineState'
 import { World } from '../ecs/classes/World'
+import { getComponent } from '../ecs/functions/ComponentFunctions'
+import { GroupComponent } from '../scene/components/GroupComponent'
 import { defaultPostProcessingSchema } from '../scene/constants/PostProcessing'
 import { createWebXRManager, WebXRManager } from '../xr/WebXRManager'
 import { XRState } from '../xr/XRState'
@@ -360,6 +364,17 @@ export default async function WebGLRendererSystem(world: World) {
     return null
   })
 
+  /** override Skeleton.update, as it is called inside  */
+  const skeletonUpdate = Skeleton.prototype.update
+
+  function noop() {}
+
+  function iterateSkeletons(skinnedMesh: SkinnedMesh) {
+    if (skinnedMesh.isSkinnedMesh) {
+      skinnedMesh.skeleton.update()
+    }
+  }
+
   const execute = () => {
     for (const action of setQualityLevelActions()) EngineRendererReceptor.setQualityLevel(action)
     for (const action of setAutomaticActions()) EngineRendererReceptor.setAutomatic(action)
@@ -370,6 +385,21 @@ export default async function WebGLRendererSystem(world: World) {
     for (const action of changeNodeHelperVisibilityActions()) EngineRendererReceptor.changeNodeHelperVisibility(action)
     for (const action of changeGridToolHeightActions()) EngineRendererReceptor.changeGridToolHeight(action)
     for (const action of changeGridToolVisibilityActions()) EngineRendererReceptor.changeGridToolVisibility(action)
+
+    /** for HMDs, only iterate priority queue entities to reduce matrix updates per frame. otherwise, this will be automatically run by threejs */
+    /** @todo include in auto performance scaling metrics */
+    if (isHMD) {
+      /**
+       * Update threejs skeleton manually
+       *  - overrides default behaviour in WebGLRenderer.render, calculating mat4 multiplcation
+       */
+      Skeleton.prototype.update = skeletonUpdate
+      for (const entity of world.priorityAvatarEntities) {
+        const group = getComponent(entity, GroupComponent)
+        for (const obj of group) obj.traverse(iterateSkeletons)
+      }
+      Skeleton.prototype.update = noop
+    }
 
     EngineRenderer.instance.execute(world.deltaSeconds)
   }
@@ -385,6 +415,7 @@ export default async function WebGLRendererSystem(world: World) {
     removeActionQueue(changeGridToolHeightActions)
     removeActionQueue(changeGridToolVisibilityActions)
     reactor.stop()
+    Skeleton.prototype.update = skeletonUpdate
   }
 
   return { execute, cleanup }
